refactor: split kompilers.py into compilers/ and transformers/ packages
kompilers.py was a monolithic proof-of-concept.  Split into:

  mvp/compilers/transformers/base.py
      Transformer base class, Handler/Matcher types, and helpers
      (_localname, _tag_matcher, _tag_attr_matcher, _rescue_tail)

  mvp/compilers/transformers/prose_transformer.py
      Family1ProseTransformer — the entry point for schema-specific
      customization by consumers (e.g. Charles Pletcher)

  mvp/compilers/transformers/registry.py
      SchemaRegistry, TransformerFactory, _default_registry

  mvp/compilers/transformers/__init__.py
      Re-exports the four public transformer symbols

  mvp/compilers/protopage_compiler.py
      ProtopageChunk, ProtopageCompiler, _sub helper

  mvp/compilers/__init__.py
      Re-exports the full public API; canonical import path is
      `from mvp.compilers import ...`

kompilers.py deleted.  test_kompilers.py updated to import from
mvp.compilers.  532 tests pass, 3 pre-existing failures unchanged.
