+++ src/python/mvp/compilers/__init__.py
@@ -1,5 +1,16 @@
-# mvp/compilers — Tools to compile new stuff from TEI documents
+from mvp.compilers.protopage_compiler import ProtopageChunk, ProtopageCompiler
+from mvp.compilers.transformers import (
+    Family1ProseTransformer,
+    SchemaRegistry,
+    Transformer,
+    TransformerFactory,
+)
 
-# from mvp.compilers.transformers import Transformer
-
-# __all__ = ["Transformer"]
+__all__ = [
+    "Family1ProseTransformer",
+    "ProtopageChunk",
+    "ProtopageCompiler",
+    "SchemaRegistry",
+    "Transformer",
+    "TransformerFactory",
+]
